Checkpointed Snapshot of Exploration, Feature Engineering and Model Selection
Data exploration, feature engineering and model selection are done on the small data set. A Fork of this file is used to create the final model from the full size data set to generate a model for inference in a web application, and create some overall statistics on the full data set.